linux-sdk: Changes to adhere to notification mechanism
Notification handles control what mode the ring will internally operate in. In this
scheme the ring will start in HID mode. If any OpenSpatial data is needed, then we need
to enable them through the corresponding CCC handles. When the connection terminates, we
need to put the ring back to the HID mode.

TEST: Pointer, free-pointer and pose6d+buttons data verified

Change-Id: Ie19cec8f3df1717bef9921c192a92cea399124da
Signed-off-by: Subash Patel <subash@enimai.com>